haha
Refactor product services and update namespaces

Significant refactor and cleanup of the `ShopHeo` application:
- Removed `ManageProductService` and `PublicProductService` classes and interfaces.
- Introduced `ProductService` to consolidate product management functionality.
- Updated dependency injection in `Startup.cs` to use `IProductService`.
- Cleaned up using directives across multiple files.
- Changed namespaces for `PagingGetManagerProductBase`, `PagingGetPublicProductBase`, and `GetUserPagingRequest` to `ShopHeo.ViewModels.Commom`.
- Updated `IUserService` and `UserService` to reflect namespace changes.
- Simplified `UserViewModel` by removing `Display` attributes.
- Added new projects to `ShopHeo.Webapp.sln` and organized them under logical folders.
- Added `Index.cshtml` view for listing users.
- Introduced `RequestBase` class with `BearerToken` property.
