+++ ShopHeo.Application/CataLog/Products/ProductService.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using ShopHeo.Application.Dtos;
 using ShopHeo.Data.EF;
 using ShopHeo.Data.Entities;
 using System;
@@ -16,15 +15,16 @@ using ShopHeo.Untitiles.Exceptions;
 using ShopHeo.ViewModels.CataLog.Products;
 using ShopHeo.ViewModels.CataLog.ProductsImage;
 using static System.Net.Mime.MediaTypeNames;
+using ShopHeo.ViewModels.Commom;
 
 namespace ShopHeo.Application.Catalog.Products
 {
-    public class ManageProductService : IManageProductService
+    public class ProductService : IProductService
     {
         private readonly string USER_CONTENT_FOLDER_NAME = "user-content";
         private readonly IStorageService storageService;
         private readonly HshopDBContext context;
-        public ManageProductService(HshopDBContext context, IStorageService storageService)
+        public ProductService(HshopDBContext context, IStorageService storageService)
         {
             this.context = context;
             this.storageService = storageService;
@@ -346,6 +346,53 @@ namespace ShopHeo.Application.Catalog.Products
                 SortOrder = image.SortOrder
             };
             return viewModel;
+        }
+
+        public async Task<PageResult<ProductViewModel>> GetAllByCategoryId(string languageId, PagingGetPublicProductBase request)
+        {
+            //1. Select join
+            var query = from p in this.context.Products
+                        join pt in this.context.ProductTranslations on p.Id equals pt.ProductId
+                        join pic in this.context.ProductInCategories on p.Id equals pic.ProductId
+                        join c in this.context.Categories on pic.CategoryId equals c.Id
+                        where pt.LanguageId == languageId
+                        select new { p, pt, pic };
+            //2. filter
+            if (request.CategoryId > 0)
+            {
+                query = query.Where(x => x.pic.CategoryId == request.CategoryId);
+            }
+            //3. Paging
+            int totalRow = await query.CountAsync();
+
+            var data = await query.Skip((request.pageIndex - 1) * request.pageSize)
+                .Take(request.pageSize)
+                .Select(x => new ProductViewModel()
+                {
+                    Id = x.p.Id,
+                    Name = x.pt.Name,
+                    DateCreated = x.p.DateCreated,
+                    Description = x.pt.Description,
+                    Details = x.pt.Details,
+                    LanguageId = x.pt.LanguageId,
+                    OriginalPrice = x.p.OriginalPrice,
+                    Price = x.p.Price,
+                    SeoAlias = x.pt.SeoAlias,
+                    SeoDescription = x.pt.SeoDescription,
+                    SeoTitle = x.pt.SeoTitle,
+                    Stock = x.p.Stock,
+                    ViewCount = x.p.ViewCount,
+                }).ToListAsync();
+
+            //4. Select and projection
+            var pagedResult = new PageResult<ProductViewModel>()
+            {
+                TotalRecords = totalRow,
+                PageSize = request.pageSize,
+                PageIndex = request.pageIndex,
+                Items = data
+            };
+            return pagedResult;
         }
     }
 }
