+++ ShopHeo.ViewModels/CataLog/Products/PagingGetManagerProductBase.cs
@@ -1,12 +1,13 @@
-﻿using System;
+﻿using ShopHeo.ViewModels.Commom;
+using System;
 using System.Collections.Generic;
 using System.Text;
-using ShopHeo.Application.Dtos;
 
 namespace ShopHeo.ViewModels.CataLog.Products
 {
     public class PagingGetManagerProductBase : PagingRequestBase
     {
+
         public string Keyword { get; set; }
         public string LanguageId { get; set; }
         public int? CategoryId { get; set; }
